Display Mock Table
Here I am just adding some functional js that will display the contents of the query in a table format. I can add some css to make this look pretty.  This is only for one of the button pages currently, but I will move this to a separate function that will be called and passed the correct parameters for each page content.